# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Documentación de la API de la aplicación",
        "version": "1.0"
    },
    "paths": {
...
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Obtener todos los recursos",
                "operationId": "7633cd726de24dafbff1c0559ade67e2",
                "responses": {
                    "200": {
                        "description": "Lista de recursos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Resource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Crear un recurso",
                "operationId": "ae1560ca6061943b6da1b27ed6e18059",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateResourceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recurso creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/resources/{id}": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Obtener un recurso",
                "operationId": "1c48e8d46ffeee459fdeccd77f879645",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Resources"
                ],
                "summary": "Desactivar un recurso",
                "operationId": "edabc99f680998ad2bda300c06eb0bab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso desactivado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Resources"
                ],
                "summary": "Actualizar un recurso",
                "operationId": "7d700fbab5332312822c1c7771517c69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateResourceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/resources/{id}/availability": {
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Verificar disponibilidad de un recurso",
                "operationId": "8da5f6a2199bb38f9792e1e9a940201d",
                "parameters": [
                    {
                        "name": "resourceId",
                        "in": "path",
                        "description": "ID del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckAvailabilityRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso disponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "available": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Recurso no disponible",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "available": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CheckAvailabilityRequest": {
                "title": "Check Availability Request",
                "description": "Solicitud para verificar la disponibilidad de una reserva",
                "required": [
                    "reserved_at",
                    "duration"
                ],
                "properties": {
                    "reserved_at": {
                        "description": "Fecha y hora de la reserva en el formato d-m-Y H:i",
                        "type": "string",
                        "format": "date-time",
                        "example": "25-12-2023 14:30"
                    },
                    "duration": {
                        "description": "Duración de la reserva en minutos, debe ser al menos 1",
                        "type": "integer",
                        "format": "int32",
                        "example": 20
                    }
                },
                "type": "object"
            },
...
            "CreateResourceRequest": {
                "title": "Create Resource Request",
                "description": "Esquema de validación para la creación de un recurso",
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del recurso",
                        "type": "string",
                        "example": "Sala de reuniones A"
                    },
                    "type": {
                        "description": "Tipo de recurso. Valores permitidos: meeting_room, equipment",
                        "type": "string",
                        "enum": [
                            "meeting_room",
                            "equipment"
                        ],
                        "example": "meeting_room"
                    },
                    "description": {
                        "description": "Descripción del recurso (opcional)",
                        "type": "string",
                        "example": "Una sala de reuniones equipada con proyector"
                    },
                    "capacity": {
                        "description": "Capacidad del recurso (opcional)",
                        "type": "integer",
                        "example": 10
                    },
                    "is_active": {
                        "description": "Estado del recurso, si está activo o no (opcional)",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "UpdateResourceRequest": {
                "title": "Update Resource Request",
                "description": "Esquema de validación para actualizar un recurso",
                "properties": {
                    "name": {
                        "description": "Nombre del recurso",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Sala de reuniones A"
                    },
                    "type": {
                        "description": "Tipo de recurso",
                        "type": "string",
                        "enum": [
                            "meeting_room",
                            "equipment"
                        ],
                        "example": "meeting_room"
                    },
                    "description": {
                        "description": "Descripción del recurso",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Una sala equipada con proyector y pizarra",
                        "nullable": true
                    },
                    "capacity": {
                        "description": "Capacidad del recurso",
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Estado de activación del recurso",
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Resource": {
                "title": "Recurso",
                "description": "Modelo que representa un recurso en el sistema de gestión de reservas.",
                "properties": {
                    "id": {
                        "description": "ID único del recurso",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del recurso",
                        "type": "string",
                        "example": "Sala de reuniones A"
                    },
                    "type": {
                        "description": "Tipo de recurso (por ejemplo, sala de reuniones o equipo)",
                        "type": "string",
                        "enum": [
                            "meeting_room",
                            "equipment"
                        ],
                        "example": "meeting_room"
                    },
                    "description": {
                        "description": "Descripción del recurso",
                        "type": "string",
                        "example": "Sala equipada con proyector y capacidad para 10 personas"
                    },
                    "capacity": {
                        "description": "Capacidad máxima del recurso",
                        "type": "integer",
                        "example": 10
                    },
                    "is_active": {
                        "description": "Indica si el recurso está activo",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del recurso",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha y hora de la última actualización del recurso",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-02T15:30:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}